+++ app/src/main/java/com/example/e_commerce/ui/lets_start_cards/Lets_start_cards.kt
@@ -1,10 +1,11 @@
-package com.example.e_commerce
+package com.example.e_commerce.ui.lets_start_cards
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.e_commerce.R
 
 class lets_start_cards : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
